+++ src/protocols/reading.rs
@@ -88,7 +88,7 @@ where
                             Err(e) => {
                                 node.known_peers().register_failure(addr);
                                 if node.config().fatal_io_errors.contains(&e.kind()) {
-                                    node.disconnect(addr);
+                                    node.disconnect(addr).await;
                                     break;
                                 } else {
                                     sleep(Duration::from_secs(